Add a main menu that iterates over dirs in "res://entries"
It is expected that the directories in "entries" will be laid out in the same way the AoC website is: YYYY/day/D/input (to get to the input for Day D in Year YYYY). Godot doesn't seem to like files without extensions, so I made the "input" actually be "input.txt".

Currently this menu checks for the existence of input.txt in the Year>Day that is selected and downloads it if it doesn't exist. This requires common/session.txt to have a valid session in it from the AoC cookie.

Next step will be to make it load up the scene for that day as well.