Add a DayMenu control
It scans the directory of the scene it is in for .txt files and has a Run button which gives the option of running the Day's solution on any of those .txt files.

The idea being that you'd create a sample.txt (as shown in this commit) to run against while building your solution, and then you'll be able to switch to the input.txt which was downloaded for you once you're ready.

Next I want to either refactor the common parts of the MainMenu and DayMenu to get rid of the duplicated bits there, or continue on with doing a solution using 2022/day/1 as a guide, and then pull out whatever boilerplate I can there to make making a Day Scene as easy as I can.