Merge Day and Main Menu into one Scene
This allowed me to make a Nav Stack for navigating through all the things. Hitting the Run button will be what actually instantiates a new Scene for Days that need that. Otherwise I should be able to do it by just adding UI via this scene and the Solution can just be a gd file